removed test messages and fixed int back to double

+++ SimpleCalculator/InputConverter.cs
@@ -4,10 +4,9 @@ namespace SimpleCalculator
 {
     public class InputConverter 
     {
-        public static int ConvertInputToNumeric(string argTextInput)
+        public static double ConvertInputToNumeric(string argTextInput)
         {
-            //testing push
-            //testing push again
+            
             int converted = Convert.ToInt32(argTextInput);
             return converted;
         }
